abstract Paywall CTAs
